+++ config.go
@@ -5,14 +5,14 @@ import (
 )
 
 type Config struct {
-	Lifetime       uint `json:"lifetime"`
-	MaxShardSize   uint `json:"max_shard_size"`
-	NumberOfShards int  `json:"number_of_shards"`
+	Lifetime       int64 `json:"lifetime"`
+	MaxShardSize   uint  `json:"max_shard_size"`
+	NumberOfShards int   `json:"number_of_shards"`
 }
 
 func DefaultConfig() Config {
 	return Config{
-		Lifetime:       uint(time.Hour * 24),
+		Lifetime:       int64(time.Hour * 24),
 		MaxShardSize:   1024,
 		NumberOfShards: 16,
 	}
